fix: Fix "Create Excalidraw Image" silently failing when the target folder does not exist
The previous attempt of determining the path, where the new excalidraw images would be created, led to errors. Now the path is determined based on the actual location of the active file, not based on the workspace folder. A new setting allows to customize, where the image will be stored. If a path doesn't exist, it will be created.